Refactor sensor system to use composition over inheritance
- New Sensor class in server/src/Sensor/Sensor.hpp uses composition
- Source hierarchy: SourceFile (reads from files), SourcePush (receives
  values from device backends)
- Transform hierarchy: TransformScale (divides by divider),
  TransformDelta
  (calculates rates from counter differences)
- Old sensor types (ValueSensor, DeltaSensor, etc.) are removed
- Device backends now use the new push-based sensor API
- Updated all device implementations to use the new sensor system
- Tests updated to use new sensor API

- Device backends now use the new push-based sensor API
- Updated all device implementations to use the new sensor system
- Tests updated to use new sensor API

This refactor makes it easier to add new sensor types and data sources
without modifying the core Sensor class.

+++ server/src/Sensor/SensorReading.hpp
@@ -5,7 +5,7 @@ struct SensorReading {
   double value;
   double min_value;
   double max_value;
-  long double sum;
+  double sum;
   size_t times;
 
   nlohmann::json serialize();
